Move lightbox to components folder
* So as not to build the lightbox as a separate page
